Add facility for overriding `HAVE_BOOST' from Dune.
This is, at best, a work-around for an issue that presents when
earlier versions of Autoconf (e.g., 2.59) are used to process Dune's
DUNE_BOOST_BASE macro (introduced in dune-common@6480,
dune-istl@1492).  For reasons I've been unable to determine, the
HAVE_BOOST symbol just gets #define-d, not set to `ENABLE_BOOST' as
was intended.  Some kind of race condition?

Anyway, the Dune ISTL module uses statements such as

   #if HAVE_BOOST
     /* ... */
   #endif

to determine if particular software features should be enabled.
This breaks down if `HAVE_BOOST' is simply defined.  Following
commit fb614100, the symbol `OPM_HAVE_BOOST' is always defined (as a
Boolean state), so if we override ISTL's notion of `HAVE_BOOST', the
above construction continues to work.

We will, however, happily remove thise kluge if a better solution
arises in the core Dune modules.