make GCC 4.4 happy
there was a "typename" outside of a template, but because the
syntactic sugar is created by the Boost unit testing framework, it was
not clear wheter it was a template or not.

Also, some harmless warnings have been fixed. These only appeared
because a few variables have recently been converted from int to
size_t...