add the custom keywords "GCOMPIDX" and "OCOMPIDX"
the intention of it is to specify the component index which should be
used to calculate the properties of the gas and oil phase in the case of
thermal black-oil runs. The keywords are supposed to use Eclipse-style
indices, i.e., "1" is stands for the first component.

The reason why I think that this is required at all is that I have not
found out if Eclipse does use temperature dependent gas viscosities
and densities for non-compositional runs. (I have found a few
equations in the RM and TD for gas which include temperature for
compositional runs on E300, though.)

Note that there won't be a WCOMPIDX because compositional simulations
in Eclipse do not seem to treat water as a separate component...

if somebody has a better idea or knows how eclipse does this for
thermal black-oil simulations, let me know...