support federated builds
i.e., with this the CMake modules do not need to be "centrally
collected" anymore, but can be shipped with the OPM modules which need
them.

This works by adding each existing "${MODULE_ROOT}/cmake/Modules"
directory for each specified variable that ends with "_ROOT" to the
CMake search path.