[cmake] Add CMake module for finding PT-Scotch in dune-cornerpoint
There already is a module for finding Zoltan, which supports
both ParMETIS and PT-Scotch underneath. While the former is
only free for academic use the latter is true open source. Therefore
with this commit we add a module for finding PT-Scotch and add the
necessary libraries and include paths for ZOLTAN.