Remove FindPetsc.cmake in favor of FindPETSc.cmake.
In PR OPM/opm-core#777 FindPetsc.cmake was removed and new
better version was established with FindPETSc.cmake. I assume
that this change was not properly propagated to all the all the
modules and some were still using FindPetsc.cmake when creating
opm-cmake. Thus the old version ended up in it.
This commit now removes it here, too.