opm_add_test: make default working directory configurable
The "default default" is the source directory of the project in order
to avoid having to copy the data files which are necessary for
executing the tests. (copying them is an almost impossible thing to do
in general, since the file system layout of the project is not
standardized.) this can be changed using the new
opm_set_test_default_working_directory() macro.

having said this, some tests may have to write some output to disk
during execution and the source directory should not be assumed to be
writable (e.g., it could be burned on a DVD), i.e., using the build
directory by default would be the more sensible choice IMO.