FindParMETIS: use the FindMETIS module to deal with sequential METIS
for me, 4292de413f56bb0511b26b7ee8412db88a9919de lead to linker errors
if the sequential metis library was required (and the parallel is not
installed).  see e.g.,

http://www.opm-project.org/CDash/viewBuildError.php?buildid=41290

The reason for this is a partial clash in the variable names which are
used by both modules. With this PR, the FindParMETIS module does not
write to any variables that are prefixed by "METIS_".