Use full path to auto-detected sources
This is a work-around for an issue that presented when switching
module opm-material's test framework to using the opm_add_test macro
with automatically detected sources.  The macro would not find any
source files and subsequently end up effectively calling

    add_executable("${CURTEST_EXE_NAME}")

which promts CMake to respond with "incorrect number of parameters".

There may be other, more fundamental, problems here, but this does
at least restore the build of module opm-material on the Jenkins CI
system.