Renamed OpmLoadDeck -> opmi
Have renamed the OpmLoadDeck test program to the working name opmi - Opm
Inspector. Currently the opmi executable will only load a deck and
create a EclipseState instance, but the plan is to extend this
executable to be a general Eclipse deck inspector, e.g.

   opmi --wells CASE.DATA

Could the list the wells and so on.