make the unit tests for the flash solvers bite if they detect an error
so far, they only printed a warning. in this case the error message
was usually ignored and the test was counted as 'passed'. (before the
switch of the flash solvers to use automatic differentiation, this
happened in some cases where single precision floating point scalars
were used.)

not anymore: now an exception is thrown and the test is aborted if an
error is detected.