Remove the OutputWriter-Eclipse inheritance
This is-a relationship is never used, and the vtable is never leveraged.
The augmented output writers already use uniquely named output writers
anyway, so there is no reason to go via the hurdle of inheritance.

This gives a lot of flexibility when it comes to interface, since
Eclipse output can make Eclipse-specific assumptions.