EclipseWriter: write restart with double.
In restart files generated by Eclipse the solution fields, i.e. pressure
and saturations are written in 32 bits precision using floats. For
compatibility that is the default behavior in opm/flow as well, but with
this patch it is optionally possible to write the solution fields with
64 bit double precision.

The resulting files will probably be incompatible with third party
applications, so this should be done with extreme caution.