OPM_THROW does not log the error message new OPM_THROW_LOG does
Logging each exception.message() is suprising for developers as
throwing an exception can often be handled by surrounding code and
does not necessarily present an error of the program. Additionally,
there was no support to suppress any logging at this point.

With this commit OPM_THROW does not log the error message any more.
For people insisting on the logging there are new macros OPM_THROW_LOG
OPM_ERROR_LOG_IF that use the old behaviour.

Still it seems more clean if the calling code does logging if it assumes
that it is needed. This might prevent multipl output in parallel programs
(as was/is the case for non-converging linear solvers.)