changed: update petsc module for newer petsc
newer petsc changed layout a bit. as of petsc 3.6.4 a
proper pkg-config file is ready for use.

additionally, align the find module with how petsc is expected to be
configured. the environment variable PETSC_DIR is the root of the source
tree and PETSC_ARCH is the build configuration, located in
$PETSC_DIR/$PETSC_ARCH

the old approach is kept for backwards compatibility but will not work
with newer petsc, in particular not if petsc is built static.