workaround cmake issue:
(older) cmake behaves unpredictable in relation to rpaths
and installation to non-system prefixes.

since some of the modules build shared (ert, opm-parser),
it's imperative that the rpath (or alternatively LD_LIBRARY_PATH) is set
in the binaries. make sure this behaves predictable by
forcing the rpath in the jenkins scripts.

in particular this broke regression tests in opm-simulators
using the comparison binaries from opm-output, as the binaries
had their rpath stripped