DEBUG keyword
Adding the DEBUG keyword specification. Makes us no longer choke on it,
but we're not using it for anything in partcular. The keyword name is
DEBUG_ since DEBUG often is defined as a macro in C++ and the auto
generation would then fail. We look up the name itself by using the
deck_names field.