make the "sortInputs" argument of the 1D function tabulation classes default to true
"1D function tabulation classes" means 'Opm::Spline' and
'Opm::Tabulated1DFunction'.

note that 'sortInputs' is a slight misnomer because these setter
methods still expect the arguments to be given in either ascending or
descending order of the 'x' values and if they are given in descending
order, the order internal copies of these arrays are reversed. a full
sort would be a bit more work because it requires to sort two
independent arrays synchronously.