black-oil: add a PVT class for solvents
this is anlogous to the dry gas PVT code but it uses the PVDS and
SDENSITY keywords.

Note that this object is only used by the eWoms black-oil solvent
code, i.e. it not used by the black oil fluid system or any other
opm-material code, but IMO opm-material is the right place for it
because of its similarity to dry gas.