Forward docstrings from C++
Some of the user facing functions and methods in sunbeam have no
implementation in the python-defined classes, but are forwarded from
boost python. This forwarding is done by hijacking the method call at
runtime, which means that python help() cannot resolve the docstring.

Copy the docstring into the hijacking attribute so that the docstring is
available as if the method/function was implemented directly in python.

help(eclipse_state_object) will now correcly forward the docstrings from
the forwarded EclipseState class.