Include <module>-prereqs.cmake in <module>-config.cmake for transitive search
Otherwise implicit dependencies are never search for. That was the case
for dune-geometry and dune-grid when configuring opm-core as these are only
transitively required via opm-grid.

With this commit we include <module>-prereqs.cmake in <module>-config.cmake
and thus trigger searches for all modules that this module depends on.

We also install the file now <module>-prereqs.cmake as we need it for
installed modules, too.