Removed dead code in from::COMPDAT
The "prev_compls" was initialized with preadded completions, such that
new records in COMPDAT would append in the correct order.

However the order is maintained without the need of including the check
in Completion::fromCOMPDAT.

Tests exist in ScheduleTests.cpp to verify completion order for both
TRACK and INPUT mode.

Completion order will affect the print in RFT data. A simple case which
will have changes for completion order was simulated and the output was
verified (with both INPUT and TRACK order).