Added implementation of serialize_SCON
The value 'Kh' (effective for connection) is not directly available from
the Connection interface.  For the time being, a dummy value is used.

Computing the correct value for 'Kh' involves using the connection
transmissibility factor (avaliable), the well diameter (available), as
well as information about grid geometry (not immediately available), well
angle through connection cell (not immediately available, as far as I
can tell), and possibly a conversion factor.