Add catch-all else clause to satisfy compiler
Some compilers (notably GCC 5.4.0; default in Ubuntu 16.04) does not
do sufficient static analysis to determine that 'vect' is actually
exhaustively defined due to all indices being enumerated, and issue
a (spurious) warning that an unassigned value may be used.

This patch provides a default branch for the if-statements so that
'vect' is always assigned to, even in the eyes of the compiler. Since
entering this branch is unequivocally a bug, an assert is added so that
the null value is not mistakenly used.