Summary: Calculate Independent Set of Vectors for Restart
This commit introduces a separate, independent SummaryState data
member (Summary::rstrt_state) that manages a small, fixed set of
summary vectors which are needed for restart purposes.

Specifically, the following set of vectors are stored/updated in

    Summary::add_timestep()

for all active wells and well groups (including FIELD):

    - Production rates for Oil, Water, Gas, and Reservoir Volume
      (i.e., (O|W|G|V)PR).

    - Cumulative production totals for Oil, Water, Gas, and
      Reservoir Volume (i.e., (O|W|G|V)PT).

    - Injection rates for Water and Gas (i.e., (W|G)IR).

    - Cumulative injection totals for Water and Gas ((W|G)IT).

    - Producing Water Cut (WCT).

    - Producing Gas/Oil ratio (GOR).

We additionally capture the well bottom-hole pressure (WBHP) for all
wells that are active at the pertinent simulation step (sim_step).

Since the set of vectors is small and fixed, we can simplify the
accumulation strategy for group and field results.  In particular,
we can start at the leaf (well) level of the (relevant) group tree
and sequentially add the well's contributions successively to each
group we encounter as we traverse the tree towards the root (field)
level.  That, in turn, means that it is sufficient to visit each
active well exactly once which reduces the number of data member
look-ups in the Schedule compared to using the prepackaged
accumulation routines (e.g., for GGPT or FOPR).

Vector naming adheres to the conventions established by LibECL,
meaning

    vector + ':' + entity

(e.g., WBHP:Inj_1 or GGOR:Prod-S) for vectors attached to wells or
groups and simply the vector name itself (e.g., FOPT) for vectors
attached to FIELD.

Add a set of unit tests to exercise the new interface.