Fix calling function from OpmPackage in package configuration file.
Find package_deps is called from <opm-module-name>-prereqs.cmake that is
included in our package configuration files. So it has to be defined early on.
We do this in the package confguration file of opm-common which needs to
be found before any other OPM module.

This fixes the CMake Error seen be e.g. dumux:
CMake Error at opm-grid/opm-grid-prereqs.cmake:37 (find_package_deps):
  Unknown CMake command "find_package_deps".